rename tic-tac-toe dir
