+++ lib/tic_tac_toe/player.rb
