--- spec/tic_tac_toe_spec.rb
@@ -1,9 +1,0 @@
-require_relative '../lib/enumerables/tic-tac-toe/game'
-require_relative '../lib/enumerables/tic-tac-toe/instructions'
-require_relative '../lib/enumerables/tic-tac-toe/main'
-require_relative '../lib/enumerables/tic-tac-toe/player'
-
-describe Game do
-  describe '#turn' do
-  end
-end
